refactor: 상속 사용
